Make the card stack peel off individual cards during a drag, not shift as one
Confirmed against the reference "Calm" launcher's own card stack
implementation (dev/barna/calm/CardStackController.kt): it recomputes
every visible card's own scale/translation/rotation/alpha continuously
from a fractional depth derived straight from live scroll position --
not by translating the whole stack by a uniform amount. The previous
live-drag mechanism (added to fix "drag doesn't respond until it
settles") did exactly the latter: a single raw pixel delta added
uniformly to every entry's already-resolved offset, so the whole stack
visibly moved together as one rigid unit instead of each card peeling
off individually.

Domain layer: CardStackLayoutPolicy.entries() gains a Float activeIndex
overload alongside the existing Int one (which now delegates to it) --
an activeIndex between two integers continuously interpolates every
entry's pose using the exact same formulas the settled case already
uses. Exact for integer inputs, so every existing caller and test is
unaffected; verified by hand against the new interpolation test's own
expected values.

App layer: CardStack.kt's CardStackInteraction gains onLiveDrag,
reporting the live raw pixel delta up to the caller (or null when idle)
instead of applying it internally. AdaptiveStageNotificationStack and
AdaptiveStageAllNotificationsStack convert that delta into a fractional
activeIndex (using the same distance threshold their own settle
already commits on, so the visual "next card is now focused" point and
the "release would commit" point coincide) and pass it into
adaptiveStageNotificationStackEntries every frame. CardStack's own
animateFloatAsState calls snap instead of animating while a drag is in
progress, since the target they'd be animating toward is already
changing every frame from the caller's live recompute.

FeedStageSurface.kt's RSS card stack shares this same CardStack engine
but isn't wired to live-drag tracking (out of scope here); removing the
buggy uniform-shift mechanism from CardStack.kt fixes it there too,
reverting it to pre-existing settle-only behavior rather than leaving
the same bug in place.

+++ app/src/main/java/com/riffle/app/launcher/rss/FeedStageSurface.kt
@@ -256,7 +256,9 @@ private fun FeedArticleStack(
                             adaptiveStageNotificationStackEntries(
                                 resolution = resolution,
                                 cardCount = cards.size,
-                                activeCardIndex = activeIndex,
+                                // Float, not (yet) live-drag-tracked here -- see the parameter's
+                                // own doc; an exact integer value behaves identically to before.
+                                activeCardIndex = activeIndex.toFloat(),
                             ),
                         animationSpec = resolution.animation,
                         reducedMotion = resolution.reducedMotion,
